split workspace tests
